Noah now receives queries, parses them and returns errors when they occur.
It is not however, doing any actions with the queries yet. But it will be easy to add more and more types of queries to noah to allow it to handle them.

The challenging part now it getting the queries to actually perform an action on the nodes in the cluster, and then returning that data efficiently.