Removed PGX dependency
Moved all PGX "pgtypes" into noah. Also moved pgio functions (although I believe cockroach already had these so i might be able to remove them later)

Added more flags for executable.

Noah's executable will serve as both a "VTGate" and a "VTTablet" depending on the "-type" flag when executing.

the "Coordinator" type will not have any query or row caching, it's too intensive to cache data for potentially so many nodes on a single machine. Data should be cached on the node leve and only for that node. This will yield the fastest queries and the largest cache.

This way we will also not have to distribute the cache if there are multiple coordinators.

I've also added a new directory called "WAL" which serves as Noah's write ahead log. I'm still trying to figure out how I want to manage this. It might be easiest to manage a "WAL" on the node level and then make it so that when query-replication is being used -> nodes need to be identical. K.I.S.S. (Keep it simple stupid)