Creating postgres driver components in noah.
Noah will need a custom driver to communicate with PostgreSQL. Since we don't want to deal with row objects or any of the other quality-of-life features of a regular driver we are creating a custom one. We will need a BAREBONE driver that can communicate with the database and maintain transactions.

I'm not sure what level of abstraction I want to have yet though.
For communicating with the workers it might be best to use gRPC since it handles a lot of the code needed for me right away.

I'll decide later.

But in regardless, I will want to support connecting directly to the postgres server without going through the abstraction layer of a sql/driver.