Finished the bare minimum for the NPGX driver. Let the testing begin.
It actually worked on my very first try without any problems. A simple SELECT 1 query, as well as the queries for the types worked perfectly. Although looking at the wire there was some weird stuff going on that I haven't seen from other drivers. So I will be investigating that; however it very well might just be the simple query request as opposed to a parse request.