Removed unnecessary variable.
Set connStatus to uninitialized until it is.

Removed connInfoFromRows for now.

I do think that if I add some custom types to PostgreSQL that the driver will then struggle to verify OIDs.
Another thing is if an OID for an extension type is static as well? What if CITEXT has OID 123 on server 1. But on server 2 it has OID 234? If that is the case I will need to build a data merging or formatting function that will unify the OIDs of types from multiple servers.
I will also need to keep track of custom types and OIDs within Noah's KV store. So that way the OIDs returned to the client are always accurate.
This will also be weird because of how I will need to handle any queries that might request OIDs from a specific server. That query could be sent to ANY server and therefore the results could vary greatly if the OIDs are not static.

IF OIDs are static though, then it will be very simple passing data between the database nodes and the client.