I've added the OID query back.
I can't find any evidence at the moment that would indicate that custom types have the same OID when created on two different servers. So I will need this query in order to properly handle data rows.

I think I will try to find a way to have Noah respond to certain OID queries with it's own values. This might get messy but I'm not sure yet how I can handle it.
I want ANY postgresql driver to be able to connect to noah and run out of the box.

Another solution might be to keep track of OIDs per connection. When a specific connection is initialized and it queries OIDs then those OIDs are cached and reused when returning other results to that connection.