I made some changes to the handling of custom types. See below
Unrecognized OIDs will now be treated as PostgreSQLs TEXT datatype. When they are transmitted to the client they will maintain this new datatype as well.

This only accounts for custom types that are similar to the TEXT datatype. like case insensitive strings.

But if the custom type were numeric or something completely different, it might be more difficult to read this data on the client-side.

With that in mind maybe in the future the client should be allowed to specify which basic data type their custom type is most similar too. This way atleast some of the data can be maintained between the systems despite differences in OIDs.

I will also need to cleanup my code here now as I've commented out a lot of things.