Tables spaces will be used for sharding.
When creating a table, the user should define a table space, "global" or "account". Global tables have the exact same data on every node in the cluster, like an accounts or users table that would need to be joined on. Account tables have tables specific to a certain "account_id"'s data. All the data for a single ID are going to be located on the same server(s). And will be distributed throughout the cluster.