Several small changes.
Health checks will only be performed by the leader of the cluster.

Setup will be performed in the cmd package.

IsLeader added to context.

When setup is run, if the node is the leader and the postgres tables
don't exist then they will be created and set as global.

Starting to make dockerfile.