Significantly improved support for multi-node transactions.
Two-phase commit is now working extremely well. Especially compared to
how it was working before... it wasn't.

I did find an error though that if an error occurs in a transaction then
no more queries can be sent over that transaction, and it must be
aborted. I will need to add some handling for this so that it will throw
out the transaction if an error happens.