Adding support for drop in functions and more types.
Drop in functions are functions that would be called in a query in
postgres, but are replaced by values from noah. For example, if a query
is sent for current_database, this query could return different results
depending on the node that serves the query, so noah will replace the
function call with a drop in value based on the context of that session
and then sent the query off to the data node. This way a lot of the
built in functions will still work consistently with their expected
behavior.